chore($web-client-config): relocate bean from web-config
- separate web-client-config
- add security-debug option
